프론트엔드 gemini 의존성 최소화
