+++ src/nest/ai/common/services/i-ai-translator-service.ts
@@ -1,5 +1,4 @@
 import { SourceLanguage } from '../../../../utils/language';
-import { AiModelName } from '../../../../ai/model';
 import { FilePathInfo } from '@/types/cache';
 
 export interface AiTranslateParam {
@@ -11,10 +10,4 @@ export interface AiTranslateParam {
   apiKey: string;
   promptPresetContent: string;
   useThinking: boolean;
-}
-
-export interface IAiTranslatorService<ModelName extends AiModelName> {
-  translate(modelName: ModelName, param: AiTranslateParam): Promise<string[]>;
-
-  getEstimatedTokenCount(texts: string[] | string): Promise<number>;
 }
