--- src/nest/ai/gemini/gemini.module.ts
@@ -1,10 +1,0 @@
-import { Module } from '@nestjs/common';
-
-import { GeminiTranslatorModule } from './translator/gemini-translator.module';
-
-@Module({
-  imports: [GeminiTranslatorModule],
-  providers: [],
-  exports: [GeminiTranslatorModule],
-})
-export class GeminiModule {}
