+++ src/types/config.d.ts
@@ -6,7 +6,6 @@ export interface TranslatorConfig {
   sourceLanguage: SourceLanguage;
   apiKey: string;
   customModelConfig: ModelConfig;
-  isCustomInputMode: boolean;
   lastPresetName?: string; // 예제 프리셋 이름
   lastPromptPresetName?: string; // 프롬프트 프리셋 이름 추가
   useThinking: boolean; // AI의 "생각" 과정 활성화 여부
